+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## 1.9.18 - 2026-03-04
+- Removed Photograph tooltip that explains copying recipe (when JEI is not installed) for the time being, because it was not working as expected.
+- Fixed log spam when trying to retrieve photograph id in some cases.
+
 ## 1.9.17 - 2026-02-27
 - Fixed **Camera Stand** not being affected by gravity.
 - [Forge] Removed dependency on the specific **Create** version.